Convert test_zzz_zzz_emmalloc_4gb to btest_exit
This simplifies the test so it can also be run under node. e.g.

```
$ ./test/runner browser.test_zzz_zzz_emmalloc_4gb --browser=node
```
